+++ mods/sbz_progression/quests/Pipeworks_and_fluid_transport.lua
@@ -81,8 +81,31 @@ Here is an example of one:
     {
         type = "quest",
         title = "Item Voids",
-        text = "Item voids delete every item that goes in, and yes these are pipeworks trashcans.",
+        text =
+        "Item voids delete every item that goes in, and yes these are pipeworks trash cans. But unlike pipeworks trash cans, they show the amount of items they've destroyed.\nThat number can \"overflow\" into the negatives, if you actually manage to do this, don't consider it a bug, but consider it an achievement :)",
         requires = { "Tubes" }
+    },
+    {
+        type = "text",
+        info = true,
+        title = "Overflow Handling",
+        requires = { "Item Voids" },
+        text = [[
+Tubes break when they have too many stacks in them. This may not appear as a problem at first, but when you think about it - it can be a huge issue.
+<b>If we have this setup:</b>
+<img name=questbook_image_basic_setup.png>
+Then, there might be an issue with it if the storinator that we are putting items to is completely filled.
+In cases where there is only one tube, the item will simply drop, but when there are at least 2 tubes, the items will wonder around, until eventually there will be too many of them. In that case, the tubes will break.
+
+How do we prevent our tubes breaking, or items dropping?
+Well... a simple answer would be to have more storinators :D... but that's not practical
+
+<b>Instead, consider using item voids like this:</b>
+<img name=questbook_image_overflow_handling.png>
+Item voids have the lowest priority, so items really don't want to go there.
+Storinator, has a higher priority than the item void, so if the storinator isn't full, items will go there.
+Item voids, have a lower priority, so if the item can't go to the storinator, it will go into the item void.
+]]
     },
     {
         type = "quest",